Outsourcing text to own .js
