+++ src/data/lang.js
@@ -1,0 +1,11 @@
+/* AboutMe */
+export const ABOUTME_NAME = `Robert Heinemann`;
+export const ABOUTME_JOB = `Frontend Web Developer`;
+export const ABOUTME_TEXT = `I am a dedicated and versatile full stack developer with a passion for creating efficient and user-friendly web applications. With 5 years of professional experience, I have worked with a variety of technologies, including React, Next.js, Node.js, MySQL, PostgreSQL, and MongoDB. My journey in web development began with a deep curiosity for how things work, and it has evolved into a career where I continuously strive to learn and adapt to new challenges. I thrive in collaborative environments and enjoy solving complex problems to deliver high-quality solutions. Outside of coding, I enjoy staying active, exploring new technologies, and contributing to open-source projects.`;
+
+/* Footer */
+export const FOOTER = {
+  address: "767 Fifth Avenue, New York, NY 10153 ",
+  phoneNo: "+12 4555 666 00 ",
+  email: "me@example.com",
+};
